+++ model/pdata/common.go
@@ -363,6 +363,39 @@ func (a AttributeValue) Equal(av AttributeValue) bool {
 	return false
 }
 
+// String converts an OTLP AttributeValue object of any type to its equivalent string
+// representation. This differs from StringVal which only returns a non-empty value
+// if the AttributeValueType is AttributeValueTypeString.
+func (a AttributeValue) AsString() string {
+	switch a.Type() {
+	case AttributeValueTypeNull:
+		return ""
+
+	case AttributeValueTypeString:
+		return a.StringVal()
+
+	case AttributeValueTypeBool:
+		return strconv.FormatBool(a.BoolVal())
+
+	case AttributeValueTypeDouble:
+		return strconv.FormatFloat(a.DoubleVal(), 'f', -1, 64)
+
+	case AttributeValueTypeInt:
+		return strconv.FormatInt(a.IntVal(), 10)
+
+	case AttributeValueTypeMap:
+		jsonStr, _ := json.Marshal(AttributeMapToMap(a.MapVal()))
+		return string(jsonStr)
+
+	case AttributeValueTypeArray:
+		jsonStr, _ := json.Marshal(attributeArrayToSlice(a.ArrayVal()))
+		return string(jsonStr)
+
+	default:
+		return fmt.Sprintf("<Unknown OpenTelemetry attribute value type %q>", a.Type())
+	}
+}
+
 func newAttributeKeyValueString(k string, v string) otlpcommon.KeyValue {
 	orig := otlpcommon.KeyValue{Key: k}
 	akv := AttributeValue{&orig.Value}
@@ -420,15 +453,33 @@ func NewAttributeMap() AttributeMap {
 	return AttributeMap{&orig}
 }
 
+// NewAttributeMapFromMap creates a AttributeMap with values
+// from the given map[string]AttributeValue.
+func NewAttributeMapFromMap(attrMap map[string]AttributeValue) AttributeMap {
+	if len(attrMap) == 0 {
+		kv := []otlpcommon.KeyValue(nil)
+		return AttributeMap{&kv}
+	}
+	origs := make([]otlpcommon.KeyValue, len(attrMap))
+	ix := 0
+	for k, v := range attrMap {
+		origs[ix].Key = k
+		v.copyTo(&origs[ix].Value)
+		ix++
+	}
+	return AttributeMap{&origs}
+}
+
 func newAttributeMap(orig *[]otlpcommon.KeyValue) AttributeMap {
 	return AttributeMap{orig}
 }
 
 // InitFromMap overwrites the entire AttributeMap and reconstructs the AttributeMap
-// with values from the given map[string]string.
+// with values from the given map[string]AttributeValue.
 //
 // Returns the same instance to allow nicer code like:
 //   assert.EqualValues(t, NewAttributeMap().InitFromMap(map[string]AttributeValue{...}), actual)
+// Deprecated: use NewAttributeMapFromMap instead.
 func (am AttributeMap) InitFromMap(attrMap map[string]AttributeValue) AttributeMap {
 	if len(attrMap) == 0 {
 		*am.orig = []otlpcommon.KeyValue(nil)
@@ -745,34 +796,9 @@ func (am AttributeMap) CopyTo(dest AttributeMap) {
 }
 
 // AttributeValueToString converts an OTLP AttributeValue object to its equivalent string representation
+// Deprecated: use AttributeValue's String method instead.
 func AttributeValueToString(attr AttributeValue) string {
-	switch attr.Type() {
-	case AttributeValueTypeNull:
-		return ""
-
-	case AttributeValueTypeString:
-		return attr.StringVal()
-
-	case AttributeValueTypeBool:
-		return strconv.FormatBool(attr.BoolVal())
-
-	case AttributeValueTypeDouble:
-		return strconv.FormatFloat(attr.DoubleVal(), 'f', -1, 64)
-
-	case AttributeValueTypeInt:
-		return strconv.FormatInt(attr.IntVal(), 10)
-
-	case AttributeValueTypeMap:
-		jsonStr, _ := json.Marshal(AttributeMapToMap(attr.MapVal()))
-		return string(jsonStr)
-
-	case AttributeValueTypeArray:
-		jsonStr, _ := json.Marshal(attributeArrayToSlice(attr.ArrayVal()))
-		return string(jsonStr)
-
-	default:
-		return fmt.Sprintf("<Unknown OpenTelemetry attribute value type %q>", attr.Type())
-	}
+	return attr.AsString()
 }
 
 // AttributeMapToMap converts an OTLP AttributeMap to a standard go map